[TS-347] Refactor: edit habit page refactoring
